Merged PR 160327: PLDM Don't retry on testcases that are supposed to return an error
## Why is this change being made?

PLDM functional tests are taking too much time to finish. There is a bottleneck on the error testcases in which the retry mechanism gets activated. It's important to not retry the pldmtool command in these testcases.

## What is changed?
To achieve the desired functionality, we pass `max_retries=1` in the target error testcases.

## How is the change being tested?

- [ ] No test, document update only
- [ ] Unit tests pass
- [ ] Integration tests pass
- [ ] Manual tests pass (add test specifics)
- [ ] Others, explain:

### Test Platform

- [ ] SVP
- [ ] FPGA (Add FPGA Config)
- [x] Silicon
- [ ] Others, explain:

## Notes

## Risks

Add don't retry feature to testcases that are supposed to return an error

Related work items: #3226647